moved flow types to different directory
